Change block gas limit data type to long

+++ src/main/java/org/ethereum/vm/client/Block.java
@@ -21,7 +21,7 @@ import java.math.BigInteger;
 
 public interface Block {
 
-    BigInteger getGasLimit();
+    long getGasLimit();
 
     byte[] getParentHash();
